refactor(zendesk): polish plugin architecture
- Inline Actions::* into their Plugins:: counterparts; drop the
  actions/ directory and the dead ActionCollectionDecorator plumbing
  in BaseCollection (the customizer wraps collections itself now).
- Extract FormBuilder and Messages sub-modules to keep each plugin file
  focused on registration and orchestration.
- Share Zendesk enum values via a new TicketEnums module so the schema
  and form builder reference a single source.
- CloseTicket: replace the single statuses option with orthogonal
  statuses + scopes so callers can pick solved vs closed and single
  vs bulk independently.
- CloseTicket: swap Zendesks raw "closed prevents ticket update" stack
  for a clean message: success ("was already closed") when targeting
  closed, Error ("cannot reopen to mark as solved") when targeting solved.
- Trim over-documented comments and route specs through #run.
