+++ packages/forest_admin_datasource_zendesk/lib/forest_admin_datasource_zendesk/collections/base_collection.rb
@@ -11,21 +11,6 @@ module ForestAdminDatasourceZendesk
       DATE_OPS    = [Operators::EQUAL, Operators::BEFORE, Operators::AFTER,
                      Operators::PRESENT, Operators::BLANK].freeze
 
-      # Pipe through an internal ActionCollectionDecorator so actions registered
-      # at the datasource level reuse the agent's form lifecycle for free.
-      def add_action(name, action)
-        super
-        action_runtime.add_action(name, action)
-      end
-
-      def execute(caller, name, data, filter = nil)
-        action_runtime.execute(caller, name, data, filter)
-      end
-
-      def get_form(caller, name, data = nil, filter = nil, metas = {})
-        action_runtime.get_form(caller, name, data, filter, metas)
-      end
-
       def aggregate(caller, filter, aggregation, _limit = nil)
         unless aggregation.operation == 'Count' && aggregation.field.nil? && aggregation.groups.empty?
           raise ForestAdminDatasourceToolkit::Exceptions::ForestException,
@@ -105,12 +90,6 @@ module ForestAdminDatasourceZendesk
       end
 
       private
-
-      def action_runtime
-        @action_runtime ||= ForestAdminDatasourceCustomizer::Decorators::Action::ActionCollectionDecorator.new(
-          self, datasource
-        )
-      end
 
       def sort_field_and_direction(entry)
         return [entry.field, entry.ascending] if entry.respond_to?(:field)
