+++ packages/forest_admin_datasource_zendesk/lib/forest_admin_datasource_zendesk/ticket_enums.rb
@@ -1,0 +1,8 @@
+module ForestAdminDatasourceZendesk
+  # Shared between the Ticket schema and plugins that build ticket forms.
+  module TicketEnums
+    STATUS   = %w[new open pending hold solved closed].freeze
+    PRIORITY = %w[low normal high urgent].freeze
+    TYPE     = %w[problem incident question task].freeze
+  end
+end
